requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
boto3==1.34.145
botocore==1.34.145
certifi==2024.7.4
charset-normalizer==3.3.2
cohere==5.6.1
dataclasses-json==0.6.7
exceptiongroup==1.2.2
fastavro==1.9.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
httpx-sse==0.4.0
huggingface-hub==0.24.0
idna==3.7
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.2.10
langchain-cohere==0.1.9
langchain-community==0.2.9
langchain-core==0.2.22
langchain-experimental==0.0.62
langchain-text-splitters==0.2.2
langsmith==0.1.93
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.6
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pydantic==2.8.2
pydantic_core==2.20.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tabulate==0.9.0
tenacity==8.5.0
tokenizers==0.19.1
tqdm==4.66.4
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
yarl==1.9.4
